ACMI-2: on-disk record structs were LP64, making Linux tapes unreadable by Tacview
PO wants to view both Wine- and Linux-generated ACMI in Tacview. Wine tapes work;
Linux tapes could not be parsed by any Win32 reader.

CAUSE. The ACMI record structs are #pragma pack(1) and written RAW -- acmirec.cpp
does fwrite(recp, sizeof(Record), 1, _fd) at 11 sites -- so the on-disk layout IS the
in-memory layout. Eight identifier fields were still declared `long`, which is 4 bytes
under Win32 and 8 under LP64:
    ACMIGenPositionData.uniqueID          ACMIFeatureStatusData.uniqueID
    ACMIFeaturePositionData.uniqueID      ACMIDOFData.uniqueID
    ACMIFeaturePositionData.leadUniqueID  ACMIAircraftPositionRecord.RadarTarget
    ACMISwitchData.uniqueID               ACMITextEvent.intTime
Every position record was therefore 4 bytes wider than the format specifies.

HALF-CONVERTED, same shape as the VUADDR-1 wire bug. The .vhs tape structs in
acmitape.h (ACMIEntityData, ACMIEntityPositionData, ACMIEventHeader, ACMIEventTrailer,
ACMIFeatEvent) were ALREADY int32_t -- which is why the tape HEADER read correctly
while the record stream was garbage, and why this went unnoticed. ACMITextEvent was
the one .vhs-written struct still on `long`. The remaining `long`s in acmitape.h are
runtime-only (SimTapeEntity holds a DrawableTrail*, ActiveEvent is a chain node) and
never reach disk, so they are left alone.

MEASURED, before and after, by walking a real tape under both layouts:
    pre-fix  tape, Win32 layout : DESYNC at byte 46, after 2 records
    pre-fix  tape, LP64  layout : 15331 records, 483379/491268 bytes, clean
    post-fix tape, LP64  layout : DESYNC at byte 534, after 14 records
    post-fix tape, Win32 layout : 6018 records, ends 3 bytes past EOF
The 3-byte overshoot is a truncated final record -- the harness killed the process
mid-write at a 160KiB buffer boundary -- not a layout error.

SEMANTICALLY VALID TOO, not just structurally: record types all legal (605 feature
positions, 206 aircraft positions, 3 stationary sfx, 260 switches, 4380 DOF, 561
feature status), timestamps monotonic 0..32886, aircraft positions in plausible Korea
coordinates (x 618824..633946, y 1938387..1972360) at 7-8 kft, one uniqueID.

Also hoisted <cstdint>, which sat BELOW the structs that now use int32_t.

NOT CONFIRMED IN TACVIEW. Tacview is not installed here -- only
INSTALL/Tacview187Setup.exe, no Wine prefix, no Tacview64.exe. So this proves the
bytes are Win32-shaped, not that Tacview opens them. Installing it would close that
gap.

BREAKING: previously recorded Linux tapes are LP64 and will no longer load. They were
unreadable by anything except the old build anyway.

TE rows 1-2 clean.

+++ src/acmi/src/include/acmirec.h
@@ -1,3 +1,7 @@
+// FF_LINUX (ACMI-2): needed by the fixed-width fields in the on-disk record
+// structs below. This include used to sit far below them.
+#include <cstdint>
+
 #ifdef FF_LINUX
 // FF_LINUX (ACMI-1): the PO gets many short .flt tapes per session.
 // StartRecording always opens a NEW numbered file (acmirec.cpp:141) and
@@ -69,6 +73,22 @@ enum
 ** Record structure typedefs for each type of record
 */
 
+// FF_LINUX (ACMI-2): every field below that identifies an object was `long`.
+// These structs are #pragma pack(1) and written to the tape RAW --
+// acmirec.cpp does fwrite(recp, sizeof(Record), 1, _fd) at 11 sites -- so the
+// on-disk layout IS the in-memory layout. `long` is 4 bytes under Win32 and 8
+// under LP64, which made Linux-written .flt/.vhs files unreadable by every
+// Win32 reader, Tacview included. Measured before the fix: parsing a Linux tape
+// with the Win32 layout desynced after 2 records (byte 46), while the LP64
+// layout walked 15331 records cleanly.
+//
+// The conversion was half done -- the .vhs structs in acmitape.h (ACMIEntityData,
+// ACMIEntityPositionData, ACMIEventHeader, ACMIEventTrailer, ACMIFeatEvent) were
+// already int32_t, which is why the tape HEADER read correctly while the record
+// stream did not. ACMITextEvent below is the one .vhs struct that was missed.
+// The remaining `long`s in acmitape.h are runtime-only structs (SimTapeEntity
+// holds a DrawableTrail*, ActiveEvent is a chain node) and never hit disk.
+//
 //
 // ACMIRecHeader
 // this struct is common thru all record types as a record header
@@ -89,7 +109,7 @@ typedef struct
 typedef struct
 {
     int type; // base type for creating simbase object
-    long uniqueID; // identifier of instance
+    int32_t uniqueID; // identifier of instance
     float x;
     float y;
     float z;
@@ -107,8 +127,8 @@ typedef struct
 typedef struct
 {
     int type; // base type for creating simbase object
-    long uniqueID; // identifier of instance
-    long leadUniqueID; // id of lead component (for bridges. bases etc)
+    int32_t uniqueID; // identifier of instance
+    int32_t leadUniqueID; // id of lead component (for bridges. bases etc)
     int slot; // slot number in component list
     int specialFlags;   // campaign feature flag
     float x;
@@ -126,7 +146,7 @@ typedef struct
 #pragma pack (push, pack1, 1)
 typedef struct
 {
-    long    intTime;
+    int32_t intTime;
     _TCHAR timeStr[20];
     _TCHAR msgStr[100];
 } ACMITextEvent;
@@ -140,7 +160,7 @@ typedef struct
 typedef struct
 {
     int type; // base type for creating simbase object
-    long uniqueID; // identifier of instance
+    int32_t uniqueID; // identifier of instance
     int switchNum;
     int switchVal;
     int prevSwitchVal;
@@ -154,7 +174,7 @@ typedef struct
 #pragma pack (push, pack1, 1)
 typedef struct
 {
-    long uniqueID; // identifier of instance
+    int32_t uniqueID; // identifier of instance
     int newStatus;
     int prevStatus;
 } ACMIFeatureStatusData;
@@ -168,7 +188,7 @@ typedef struct
 typedef struct
 {
     int type; // base type for creating simbase object
-    long uniqueID; // identifier of instance
+    int32_t uniqueID; // identifier of instance
     int DOFNum;
     float DOFVal;
     float prevDOFVal;
@@ -277,7 +297,7 @@ typedef struct
 {
     ACMIRecHeader hdr;
     ACMIGenPositionData data;
-    long RadarTarget;
+    int32_t RadarTarget;
 
 } ACMIAircraftPositionRecord;
